// api/src/enums/GQLMethod.ts
enum GQLMethod {
    INDIRECT_METHOD = 'INDIRECT_METHOD',
    UPDATE_PROPERTIES = 'UPDATE_PROPERTIES',
    AUTO_MERGE = 'AUTO_MERGE',
    DUPLICATE = 'DUPLICATE',
    FINALIZE_REVISION = 'FINALIZE_REVISION',
    PUBLISH_PLATFORM = 'PUBLISH_PLATFORM',
    REORDER_LINKED_ENTITIES = 'REORDER_LINKED_ENTITIES',
    DELETE_LINKED_ENTITY = 'DELETE_LINKED_ENTITY',
    ADD_LINKED_ENTITY = 'ADD_LINKED_ENTITY',
    CREATE = 'CREATE',
    DELETE = 'DELETE',
}

export default GQLMethod;